fix(liquidate): liquidate even if maxMint reached

+++ test/fuzz/Vault.t.sol
@@ -335,8 +335,8 @@ contract VaultFuzzTest is ERC4626Test, Helpers {
   }
 
   function test_liquidate(Init memory init, uint256 shares) public virtual {
-    // We set the higher bound to uint88 to avoid overflowing above uint96
-    init.yield = int(bound(shares, 10e18, type(uint88).max));
+    // We set the higher bound to uint104 to avoid overflowing above uint112
+    init.yield = int(bound(shares, 10e18, type(uint104).max));
 
     setUpVault(init);
     vault.setLiquidationPair(ILiquidationPair(address(liquidationPair)));
@@ -367,7 +367,7 @@ contract VaultFuzzTest is ERC4626Test, Helpers {
     }
   }
 
-  // We set the higher bound to uint88 to avoid overflowing above uint96
+  // We set the higher bound to uint104 to avoid overflowing above uint112
   function setUpVault(Init memory init) public virtual override {
     for (uint i = 0; i < N; i++) {
       init.user[i] = makeAddr(Strings.toString(i));
@@ -375,7 +375,7 @@ contract VaultFuzzTest is ERC4626Test, Helpers {
 
       vm.assume(_isEOA(user));
 
-      uint shares = bound(init.share[i], 0, type(uint88).max);
+      uint shares = bound(init.share[i], 0, type(uint104).max);
       try IMockERC20(_underlying_).mint(user, shares) {} catch {
         vm.assume(false);
       }
@@ -387,7 +387,7 @@ contract VaultFuzzTest is ERC4626Test, Helpers {
         vm.assume(false);
       }
 
-      uint assets = bound(init.asset[i], 0, type(uint88).max);
+      uint assets = bound(init.asset[i], 0, type(uint104).max);
       try IMockERC20(_underlying_).mint(user, assets) {} catch {
         vm.assume(false);
       }
@@ -397,22 +397,22 @@ contract VaultFuzzTest is ERC4626Test, Helpers {
   }
 
   function _max_deposit(address from) internal virtual override returns (uint) {
-    if (_unlimitedAmount) return type(uint88).max;
-    return uint88(IERC20(_underlying_).balanceOf(from));
+    if (_unlimitedAmount) return type(uint104).max;
+    return uint104(IERC20(_underlying_).balanceOf(from));
   }
 
   function _max_mint(address from) internal virtual override returns (uint) {
-    if (_unlimitedAmount) return type(uint96).max;
-    return uint88(vault_convertToShares(IERC20(_underlying_).balanceOf(from)));
+    if (_unlimitedAmount) return type(uint112).max;
+    return uint104(vault_convertToShares(IERC20(_underlying_).balanceOf(from)));
   }
 
   function _max_withdraw(address from) internal virtual override returns (uint) {
-    if (_unlimitedAmount) return type(uint88).max;
-    return uint88(vault_convertToAssets(IERC20(_vault_).balanceOf(from)));
+    if (_unlimitedAmount) return type(uint104).max;
+    return uint104(vault_convertToAssets(IERC20(_vault_).balanceOf(from)));
   }
 
   function _max_redeem(address from) internal virtual override returns (uint) {
-    if (_unlimitedAmount) return type(uint88).max;
-    return uint88(IERC20(_vault_).balanceOf(from));
+    if (_unlimitedAmount) return type(uint104).max;
+    return uint104(IERC20(_vault_).balanceOf(from));
   }
 }
